Fixed bugs.
Equip drag widget did not show all equipped items.
When character in combat - auto hp&mp regen are stopped correctly.

Added an Enum which represents the equipment order in which the equipment arrays, on the equipment widget and equipment drag widget, are ordered.
Also made the code shorter & more readable.