Added new textures, fixed bug regarding screen focus and stats
fixed a bug where stats could not be added if a player tried to add and had 0 AP available.
Also fixed visual bug in the stats widget which shows stats, ap, etc...
Fixed a bug where at game start, needed to click with the mouse on screen in order to move.
Added new two backgrounds (need to change to a higher resolution but will do for now)