Added more implementations using the base values from playerstate cpp class
Relying less and less on charBP to handle all the player state. The logic worked fine, it was just not the right place for it to be, hence moving the state of the player (also things needed to be saved on DB for persistence system) to the player state custom class.

Added more setters, functions and enum on cpp side in order to help implementing those changes.